Add test
This commit includes a new function for regress_gp and a new sql test.
The sql test creates two resource groups: one with big memory_limit to
fill the global shared memory quota, and the other with concurrency=1.
The latter is needed to test memory counting in it. Small memory_limit
is needed so that we can fill the entire memory with one allocation and
concurrency=1 is needed to simplify calculations so that one query could
consume the entire group's quota.

The new function gp_occupy_resgroup_memory in regress_gp just calculates
the total group's amount of memory and tries to allocate it. In the
patched version of gpdb this would fail because of startup memory already
consuming some part of this quota, so the overusage would equal the
amount of this memory. However, the unpatched version wouldn't fail
because startup memory is tracked incorrectly.